Second Commit

+++ Service/Nums/AcceptsTwoNums.cs
@@ -1,4 +1,4 @@
-using MC_TwoToFour___Endpoints.Service.Nums;
+
 
 namespace MC_TwoToFour___Endpoints.Service.Nums;
 public class AcceptsTwoNums : IAcceptsTwoNums
@@ -9,7 +9,7 @@ public class AcceptsTwoNums : IAcceptsTwoNums
     {
         string userFirstNum = "";
         string userSecondNum = "";
-        int userNumResult = Convert.ToInt32(userFirstNum +  userSecondNum);
+        double userNumResult = Convert.ToDouble(userFirstNum +  userSecondNum);
         
         return $"The sum of {firstNum} and {secondNum} is  {userNumResult}";
